requirements.txt
appdirs==1.4.3
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
blinker==1.4
cffi==1.14.0
click==7.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathspec==0.8.0
pycparser==2.20
pymongo==3.4
regex==2020.4.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
flask_mongoengine
pyjwt
